+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/restapi/apis/SetGraffiti.java
@@ -13,79 +13,64 @@
 
 package tech.pegasys.teku.validator.client.restapi.apis;
 
-import static tech.pegasys.teku.ethereum.json.types.SharedApiTypes.PUBKEY_API_TYPE;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_INTERNAL_SERVER_ERROR;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_NOT_FOUND;
-import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_OK;
+import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_NO_CONTENT;
+import static tech.pegasys.teku.infrastructure.json.types.CoreTypes.STRING_TYPE;
 import static tech.pegasys.teku.validator.client.restapi.ValidatorRestApi.TAG_GRAFFITI;
 import static tech.pegasys.teku.validator.client.restapi.ValidatorTypes.PARAM_PUBKEY_TYPE;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import java.nio.charset.StandardCharsets;
 import java.util.Optional;
 import java.util.function.Function;
-import org.apache.tuweni.bytes.Bytes;
-import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.infrastructure.json.types.DeserializableTypeDefinition;
-import tech.pegasys.teku.infrastructure.json.types.SerializableTypeDefinition;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.EndpointMetadata;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiEndpoint;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiRequest;
-import tech.pegasys.teku.validator.api.Bytes32Parser;
 import tech.pegasys.teku.validator.api.GraffitiManagementException;
-import tech.pegasys.teku.validator.api.UpdatableGraffitiProvider;
+import tech.pegasys.teku.validator.api.GraffitiManager;
 import tech.pegasys.teku.validator.client.KeyManager;
 import tech.pegasys.teku.validator.client.Validator;
 
-public class GetGraffiti extends RestApiEndpoint {
-  static final String ROUTE = "/eth/v1/validator/{pubkey}/graffiti";
+public class SetGraffiti extends RestApiEndpoint {
   private final KeyManager keyManager;
+  private final GraffitiManager graffitiManager;
 
-  public static final DeserializableTypeDefinition<Bytes32> GRAFFITI_TYPE =
-      DeserializableTypeDefinition.string(Bytes32.class)
-          .formatter(GetGraffiti::processGraffitiString)
-          .parser(Bytes32Parser::toBytes32)
-          .example("Example graffiti")
-          .description("Bytes32 string")
-          .format("byte")
-          .build();
-
-  private static final SerializableTypeDefinition<Optional<Bytes32>> GRAFFITI_RESPONSE_TYPE =
-      SerializableTypeDefinition.<Optional<Bytes32>>object()
-          .withOptionalField("pubkey", PUBKEY_API_TYPE, value -> Optional.empty())
+  public static final DeserializableTypeDefinition<String> GRAFFITI_REQUEST_TYPE =
+      DeserializableTypeDefinition.object(String.class, StringBuilder.class)
+          .initializer(StringBuilder::new)
+          .finisher(StringBuilder::toString)
           .withField(
               "graffiti",
-              GRAFFITI_TYPE,
-              graffiti -> graffiti.orElse(Bytes32Parser.toBytes32(Bytes.EMPTY.toArray())))
+              STRING_TYPE.withDescription(
+                  "Arbitrary data to set in the graffiti field of BeaconBlockBody"),
+              Function.identity(),
+              StringBuilder::append)
           .build();
 
-  private static final SerializableTypeDefinition<Optional<Bytes32>> RESPONSE_TYPE =
-      SerializableTypeDefinition.<Optional<Bytes32>>object()
-          .name("GraffitiResponse")
-          .withField("data", GRAFFITI_RESPONSE_TYPE, Function.identity())
-          .build();
-
-  public GetGraffiti(final KeyManager keyManager) {
+  public SetGraffiti(final KeyManager keyManager, final GraffitiManager graffitiManager) {
     super(
-        EndpointMetadata.get(ROUTE)
-            .operationId("getGraffiti")
-            .summary("Get Graffiti")
-            .description(
-                "Get the graffiti for an individual validator. If no graffiti is set explicitly, returns the process-wide default.")
+        EndpointMetadata.post(GetGraffiti.ROUTE)
+            .operationId("setGraffiti")
+            .summary("Set Graffiti")
+            .description("Set the graffiti for an individual validator.")
             .tags(TAG_GRAFFITI)
             .withBearerAuthSecurity()
             .pathParam(PARAM_PUBKEY_TYPE)
-            .response(SC_OK, "Success response", RESPONSE_TYPE)
+            .requestBodyType(GRAFFITI_REQUEST_TYPE)
+            .response(SC_NO_CONTENT, "Successfully updated graffiti.")
             .withAuthenticationResponses()
             .withNotFoundResponse()
             .build());
     this.keyManager = keyManager;
+    this.graffitiManager = graffitiManager;
   }
 
   @Override
   public void handleRequest(final RestApiRequest request) throws JsonProcessingException {
     final BLSPublicKey publicKey = request.getPathParameter(PARAM_PUBKEY_TYPE);
+    final String graffiti = request.getRequestBody();
 
     final Optional<Validator> maybeValidator = keyManager.getValidatorByPublicKey(publicKey);
     if (maybeValidator.isEmpty()) {
@@ -94,15 +79,10 @@ public class GetGraffiti extends RestApiEndpoint {
     }
 
     try {
-      final UpdatableGraffitiProvider provider =
-          (UpdatableGraffitiProvider) maybeValidator.get().getGraffitiProvider();
-      request.respondOk(provider.getUnsafe());
+      graffitiManager.setGraffiti(publicKey, graffiti);
+      request.respondWithCode(SC_NO_CONTENT);
     } catch (GraffitiManagementException e) {
       request.respondError(SC_INTERNAL_SERVER_ERROR, e.getMessage());
     }
-  }
-
-  private static String processGraffitiString(final Bytes32 graffiti) {
-    return new String(graffiti.toArrayUnsafe(), StandardCharsets.UTF_8).strip().replace("\0", "");
   }
 }